data-updater/requirement.txt
requests==2.23.0
pyyaml==5.3.1
psycopg2-binary==2.8.4